Passing in wash services more generically
The constructor for CarJobProcessorService is getting bloated. If the user creates more wash services, they will need to add the class, add it to the correct enum, then add it to the constructor of the processor service.
Added 2 new interfaces. IWashType and IAddOnType. This moves the responsibility of identifying the type to the class. All wash services are then sent together in the constructor, where the list will be separated into two dictionaries based on being a wash or an add on.
Now, the user only needs to add to the proper enumerator and create the class. No additional work necessary!